Add syntax migration messages to localization file
Changes were made to the way syntaxes are handled in the SyntaxLoader. Specifically, a message about syntax migration was added to the English.xaml localization file. This message is intended to inform the user about changes to syntax handling (such as deleted old syntax files and the addition of default syntax highlighting). Also, redundant logging was removed from the syntax download error handling section.